Changes to be committed: 	modified:   cal.html 	modified:   js/cal.js 	modified:   js/logic.js 	modified:   js/profile.js 	modified:   js/referendum.js 	modified:   js/scheda_elettorale.js 	modified:   opt.html 	modified:   profilo.html 	modified:   referendum.html 	modified:   scheda_elettorale.html

+++ js/logic.js
@@ -241,8 +241,10 @@ function log_in(){
 }
 
 function log_out(){
-	localStorage.setItem('logged_user', JSON.stringify(null));
-	window.location.href = "index.html";
+	if(window.confirm("Vuoi disconnetterti?")){
+		localStorage.setItem('logged_user', JSON.stringify(null));
+		window.location.href = "index.html"
+	}
 }
 
 function register(){
